WIP: move handlers and case
