+++ internal/server/http/user/handlers.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/ITA-Dnipro/Dp-210_Go/internal/entity"
 	"github.com/ITA-Dnipro/Dp-210_Go/internal/server/http/customerrors"
-	md "github.com/ITA-Dnipro/Dp-210_Go/internal/server/http/middleware"
+	"github.com/ITA-Dnipro/Dp-210_Go/internal/server/http/middleware"
 
 	authPkg "github.com/ITA-Dnipro/Dp-210_Go/internal/service/auth"
 	"github.com/go-chi/chi"
@@ -56,45 +56,32 @@ func NewHandlers(uc UsersUsecases, log *zap.Logger, auth Auth) *Handlers {
 	return &Handlers{userCases: uc, logger: log, auth: auth}
 }
 
-// GetToken by basic auth.
-func (h *Handlers) GetToken(w http.ResponseWriter, r *http.Request) {
-	var newUser entity.NewUser
-	if err := json.NewDecoder(r.Body).Decode(&newUser); err != nil {
-		h.writeErrorResponse(http.StatusBadRequest, "can't parse a user", w)
+func (h *Handlers) ChangePassword(w http.ResponseWriter, r *http.Request) {
+	var req entity.UserNewPassword
+	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
+		h.writeErrorResponse(http.StatusBadRequest, "could not parse request", w)
 		return
 	}
-	if ok := isRequestValid(&newUser); !ok {
-		h.writeErrorResponse(http.StatusBadRequest, "user data invalid", w)
-		return
-	}
-	user, err := h.userCases.Authenticate(r.Context(), newUser.Email, newUser.Password)
-	if err != nil {
-		h.writeErrorResponse(http.StatusUnauthorized, err.Error(), w)
-		return
-	}
-	var tkn struct {
-		Token authPkg.JwtToken `json:"token"`
-	}
-	tkn.Token, err = h.auth.CreateToken(authPkg.UserAuth{Id: user.ID, Role: user.PermissionRole})
-	if err != nil {
-		h.writeErrorResponse(http.StatusUnauthorized, err.Error(), w)
-		return
-	}
-	h.logger.Info("ger all request succeeded")
-	h.render(w, tkn)
-}
 
-func (h *Handlers) LogOut(w http.ResponseWriter, r *http.Request) {
-	u, ok := md.UserFromContext(r.Context())
+	u, ok := middleware.UserFromContext(r.Context())
 	if !ok {
-		h.writeErrorResponse(http.StatusUnauthorized, "no such session", w)
+		h.writeErrorResponse(http.StatusUnauthorized, "auth error", w)
+		return
+	}
+	req.UserID = u.Id
+
+	if req.Password == "" {
+		h.writeErrorResponse(http.StatusBadRequest, "request does not meet needed criterium", w)
 		return
 	}
 
-	if err := h.auth.InvalidateToken(u.Id); err != nil {
-		h.logger.Warn(fmt.Sprintf("log out: user %v; err: %v", u.Id, err))
-		h.writeErrorResponse(http.StatusInternalServerError, "could not log out", w)
+	if req.Password != req.PasswordConfirm {
+		h.writeErrorResponse(http.StatusBadRequest, "new password and new password confirm do not match", w)
 		return
+	}
+
+	if err := h.paswCases.ChangePassword(r.Context(), req); err != nil {
+		h.writeErrorResponse(http.StatusForbidden, "wrong password", w)
 	}
 
 	w.WriteHeader(http.StatusOK)
